refactor(allocator): extract /api/health into routes/health.py
Moves _tailscale_status and health_check out of main.py into a new
health blueprint, following the routes/desktop.py and
routes/registration.py pattern (lazy `from lablink_allocator_service
import main` for state reads, current_app.config for app config).
main.py remains the single owner of database/scheduler_service/
reboot_service/_startup_time; the blueprint reads them lazily at
request time.
