+++ packages/allocator/src/lablink_allocator_service/main.py
@@ -47,6 +47,7 @@ from lablink_allocator_service.client_session import RotationFailed
 from lablink_allocator_service.providers.registry import get_provider
 from lablink_allocator_service.secret_hash import hash_secret
 from lablink_allocator_service.routes.desktop import bp as desktop_bp
+from lablink_allocator_service.routes.health import bp as health_bp
 from lablink_allocator_service.routes.internal_proxy_auth import (
     bp as internal_proxy_auth_bp,
 )
@@ -99,6 +100,7 @@ app.wsgi_app = _ProxyFixWhenTrusted(
     app.wsgi_app, trust_headers=lambda: should_use_https(cfg)
 )
 app.register_blueprint(desktop_bp)
+app.register_blueprint(health_bp)
 app.register_blueprint(internal_proxy_auth_bp)
 app.register_blueprint(registration_bp)
 
@@ -251,63 +253,6 @@ def notify_participants():
 @app.route("/")
 def home():
     return render_template("index.html")
-
-
-def _tailscale_status() -> str:
-    """Return "ok" / "not joined" for the allocator's own tailnet
-    connection.
-
-    The mesh-overlay sidecar shares the allocator's *network* namespace
-    (network_mode: service:allocator) but not its filesystem, so the
-    `tailscale` CLI binary — which lives only in the sidecar's image —
-    is never present here; shelling out to it would always report
-    "not installed" regardless of whether the sidecar actually joined.
-
-    Checking for the shared `tailscale0` interface's existence alone is
-    not enough either: confirmed live against a real tailnet, the kernel
-    interface stays up (`UP,LOWER_UP`, with a link-local IPv6 address)
-    even while the node is logged out and unauthenticated with the
-    control plane — a control-plane hiccup can leave the device node
-    behind with no working overlay path. Requiring an actual Tailscale
-    IPv4 address (only assigned once the control plane has authenticated
-    the node) avoids that false positive, still with no CLI and no
-    socket-sharing between the two containers."""
-    try:
-        result = subprocess.run(
-            ["ip", "-4", "-o", "addr", "show", "tailscale0"],
-            capture_output=True,
-            text=True,
-            timeout=5,
-        )
-    except (OSError, subprocess.TimeoutExpired):
-        return "not joined"
-    return "ok" if result.returncode == 0 and "inet " in result.stdout else "not joined"
-
-
-@app.route("/api/health", methods=["GET"])
-def health_check():
-    """Return structured readiness status."""
-    checks = {
-        "database": "ok" if database is not None else "not initialized",
-        "scheduler": "ok" if scheduler_service is not None else "not initialized",
-        "reboot_service": "ok" if reboot_service is not None else "not initialized",
-    }
-    if app.config["LABLINK_PROVIDER"].client_connectivity.requires_tailscale_check:
-        checks["tailscale"] = _tailscale_status()
-
-    all_ready = all(v == "ok" for v in checks.values())
-    status = "healthy" if all_ready else "starting"
-    code = 200 if all_ready else 503
-
-    payload = {
-        "status": status,
-        "checks": checks,
-    }
-
-    if _startup_time is not None:
-        payload["uptime_seconds"] = round(time.monotonic() - _startup_time, 1)
-
-    return jsonify(payload), code
 
 
 @app.route("/admin/create")
